fix DI issues; create TokenManager to check refreshToken and make api call
