+++ app/src/main/java/com/example/tasky/common/domain/TokenManager.kt
@@ -1,6 +1,45 @@
 package com.example.tasky.common.domain
 
-interface TokenManager {
-    suspend fun isTokenExpired(): Boolean
-    suspend fun refreshToken()
+import android.util.Log
+import com.example.tasky.common.domain.repository.TokenRemoteRepository
+import kotlinx.coroutines.flow.first
+import java.time.ZonedDateTime
+import javax.inject.Inject
+
+class TokenManager @Inject constructor(
+    private val tokenRepository: TokenRemoteRepository,
+    private val sessionStateManager: SessionStateManager
+) {
+    companion object {
+        private val TAG = "TokenManager"
+    }
+
+    suspend fun shouldRefreshToken() {
+        val now = ZonedDateTime.now()
+        val currentTimeInMilliseconds = now.toInstant().toEpochMilli()
+        val expirationTime = sessionStateManager.getAccessTokenExpiration().first() ?: 0L
+
+        if (currentTimeInMilliseconds > expirationTime) {
+            refreshToken()
+        }
+    }
+
+    private suspend fun refreshToken() {
+        Log.d(TAG, "refreshToken()")
+        val refreshToken = sessionStateManager.getRefreshToken() ?: ""
+        val userId = sessionStateManager.getUserId().first()
+
+        val result = tokenRepository.refreshSession(refreshToken, userId.toString())
+
+        when (result) {
+            is Result.Success -> {
+                println("Success refresh token!")
+                sessionStateManager.setAccessToken(result.data.accessToken)
+            }
+
+            is Result.Error -> {
+                println("Failed refresh token :(")
+            }
+        }
+    }
 }
